move validator func to utils
