+++ packages/mirrorful/editor/src/components/ColorPalette/EditColorModal.tsx
@@ -18,10 +18,9 @@ import {
 } from '@chakra-ui/react'
 import { TColorData } from 'types'
 import { useState, useRef } from 'react'
-import { generateDefaultColorShades } from './utils'
+import { generateDefaultColorShades, handleInvalidColor } from './utils'
 import { ColorPicker } from './ColorPicker'
 import { Color } from '@hello-pangea/color-picker'
-import { VALID_CSS_COLORS } from './validCssColors'
 import tinycolor from 'tinycolor2'
 import { FaMagic } from 'react-icons/fa'
 
@@ -68,38 +67,6 @@ export function EditColorModal({
   const shouldRecommendHover = !!base && !hover
   const shouldRecommendActive = !!base && !active
   const shades = generateDefaultColorShades(base)
-
-  const handleInvalidColor = (input: string) => {
-    // Check if input is a valid hex code
-    const hexRegex = /^#?([0-9A-Fa-f]{3}|[0-9A-Fa-f]{6})$/
-    if (hexRegex.test(input)) {
-      if (input.startsWith('#')) {
-        return input
-      } else {
-        return `#${input}`
-      }
-    }
-
-    // Check if input is a valid color name
-    const lowerCaseInput = input.toLowerCase()
-    if (VALID_CSS_COLORS.includes(lowerCaseInput)) {
-      return lowerCaseInput
-    }
-
-    const validSubsetRegex = /^#?[0-9A-Fa-f]{0,6}$/ // regex to validate if input is a valid subset of a hexcode
-    const randomHex = Math.floor(Math.random() * 16777215).toString(16) // generate a random valid hexcode
-
-    if (validSubsetRegex.test(input)) {
-      // check if input is a valid subset of a hexcode
-      if (input.startsWith('#')) {
-        return `${input}${randomHex.slice(input.length - 1)}` // use input as the first part and append random characters as necessary to make a valid hexcode
-      } else {
-        return `#${input}${randomHex.slice(input.length)}` // use input as the first part and append random characters as necessary to make a valid hexcode
-      }
-    } else {
-      return `#${randomHex}` // generate a completely random valid hexcode
-    }
-  }
 
   const onBaseBlur = () => {
     const value = handleInvalidColor(base)
